+++ CHANGELOG.md
@@ -2,15 +2,18 @@
 
 ## CHANGELOG
 
-### 3.0.0 (19-Aug-2025)
+### 3.0.0 (20-Aug-2025)
 - First (attempt at a) MDM-agnostic release
 
-### 2.3.0 (19-Aug-2025)
+### 2.3.0 (20-Aug-2025)
 - Enhanced `operationMode` to verbosely execute when set to `debug` (Addresses Issue #28)
 - Adjusted GlobalProtect VPN check for IPv6
 - Enhanced `checkJssCertificateExpiration` function (Addresses Issue #27 via Pull Request #30; thanks, @theahadub and @ScottEKendall)
 - Extended Network Checks (Pull Request #31 addresses Issue #23; thanks big bunches, @tonyyo11!)
 - Added `organizationBrandingBannerURL` (thanks for the inspiration, @ScottEKendall!) [Image by benzoix on Freepik](https://www.freepik.com/author/benzoix)
+- Adjusted `checkVPN` function to report "Unknown" for the catch-all condition of `vpnStatus`
+- Added "Connected" and "Disconnected" options to `checkVPN` function
+- Adjusted Palo Alto Networks GlobalProtect VPN Information
 
 ### 2.2.0 (15-Aug-2025)
 - Improved the GlobalProtect VPN IP detection logic